+++ backend/src/routes/v0/__init__.py
@@ -10,7 +10,7 @@ from .storage import router as storage
 from .assistant import router as assistant
 from .schedule import router as schedule
 from .prompt import router as prompt
-from .prompt import static_router as static_prompt
+
 __all__ = [
     "llm",
     "thread",
@@ -22,7 +22,6 @@ __all__ = [
     "assistant",
     "schedule",
     "prompt",
-    "static_prompt",
 ]
 
 if LANGCONNECT_SERVER_URL:
